# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.NginxConfig:
    properties:
      accessLog:
        type: string
      configPath:
        type: string
      errorLog:
        type: string
      gzip:
        type: boolean
      pidFile:
        type: string
      serverTokens:
        type: boolean
      user:
        type: string
      workerConn:
        type: string
      workerProcess:
        type: string
    type: object
...
  models.NginxStatus:
    properties:
      configTest:
        type: boolean
      connections:
        type: integer
      pid:
        type: integer
      running:
        type: boolean
      uptime:
        type: string
      version:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@etapanel.com
    name: EtaPanel Support
    url: https://github.com/EtaPanel-dev/EtaPanel
  description: EtaPanel 服务器管理面板后端API文档，提供系统监控、文件管理、Nginx配置、SSL证书管理等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EtaPanel API
  version: "1.0"
paths:
...
  /auth/nginx/config:
    get:
      consumes:
      - application/json
      description: 获取Nginx的主配置文件内容
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.NginxConfig'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取Nginx主配置
      tags:
      - Nginx管理
    put:
      consumes:
      - application/json
      description: 更新Nginx的主配置文件
      parameters:
      - description: Nginx配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NginxConfig'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 更新Nginx主配置
      tags:
      - Nginx管理
  /auth/nginx/config/reset:
    post:
      consumes:
      - application/json
      description: 将Nginx配置重置为默认设置
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 重置Nginx配置
      tags:
      - Nginx管理
...
  /auth/nginx/status:
    get:
      consumes:
      - application/json
      description: 获取Nginx服务的运行状态信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.NginxStatus'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取Nginx状态
      tags:
      - Nginx管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入Bearer token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...